Fix #25 : data type issue std::time_t
std::time_t is 32 bit on the Raspberry Pi CPU and 64 on most modern
PC's. This causes a warning in the old code. The new works now correct
on both CPUs 32 bit and 64 bit. The 32 bit solution will fail in year 2106.

Background:
If time_t is 32 bit 1.1.1970 + (2^32)s = 7. Feb 2106 at 07:28:15. At that
time I will not care anymore if my code fails ;-).